+++ src/components/Page/ProgressModalPage.stories.tsx
@@ -7,7 +7,6 @@
 import React from "react";
 import { Meta, Story } from "@storybook/react/types-6-0";
 import { ProgressModalPage, ProgressModalPageProps } from "./ProgressModalPage";
-import "../../styles/icons/ab_icons.css";
 import { Button } from "..";
 
 export default {