+++ src/utils/index.ts
@@ -5,7 +5,6 @@
  */
 
 export * from "./common";
-export * from "./datetime";
 export * from "./input";
 export * from "./makeSafeReactInnerHTML";
 export * from "./password";